+++ src/documentStore.ts
@@ -6,7 +6,7 @@ import {
   watchConfigSettings,
 } from './config';
 import { DocumentStore as IDocumentStore } from './extensionApi';
-import { logToOutputChannelFactory } from './io';
+import { logToOutputChannelFactory, LogChannel, resetOutputChannel } from './io';
 import * as utils from './utils';
 import * as httpyac from 'httpyac';
 import * as vscode from 'vscode';
@@ -221,12 +221,14 @@ export class DocumentStore extends utils.DisposeProvider implements IDocumentSto
 
   async send(context: httpyac.HttpRegionSendContext | httpyac.HttpFileSendContext | undefined): Promise<boolean> {
     try {
+      resetOutputChannel();
       if (context) {
         const config = await getEnvironmentConfig(context.httpFile.fileName);
+
         if (!context.scriptConsole) {
           context.scriptConsole = new httpyac.io.Logger({
             level: config.log?.level,
-            logMethod: logToOutputChannelFactory('Console'),
+            logMethod: logToOutputChannelFactory(LogChannel.Console, LogChannel.Log),
           });
         }
         if (!context.config) {