+++ src/io/index.ts
@@ -1,11 +1,4 @@
-export { toUri } from './fileProvider';
+export * from './fileProvider';
+export * from './initIOProvider';
 export * from './storageProvider';
-export { getOutputChannel, logStream, logToOutputChannelFactory } from './userInteractionProvider';
-import { initFileProvider } from './fileProvider';
-import { initLog, initUserInteractionProvider } from './userInteractionProvider';
-
-export function initIOProvider() {
-  initFileProvider();
-  initLog();
-  return initUserInteractionProvider();
-}
+export * from './userInteractionProvider';