+++ src/io/initIOProvider.ts
@@ -1,0 +1,8 @@
+import { initFileProvider } from './fileProvider';
+import { initLog, initUserInteractionProvider } from './userInteractionProvider';
+
+export function initIOProvider() {
+  initFileProvider();
+  initLog();
+  return initUserInteractionProvider();
+}